Introduce metadataToBytes
