+++ packages/ui/src/common/model/JoystreamNode/index.ts
@@ -1,1 +1,4 @@
 export * from './getEventParam'
+export { metadataToBytes } from '@/common/model/JoystreamNode/metadataToBytes'
+export { AnyMetadataClass } from '@/common/model/JoystreamNode/metadataToBytes'
+export { AnyMessage } from '@/common/model/JoystreamNode/metadataToBytes'
